PIをlong doubleに

+++ library/def/const.hpp
@@ -8,7 +8,7 @@ const long long MAX = 2e6;
 const long long MOD = 998244353;
 const long long MOD7 = 1e9 + 7;
 const long double EPS = 1e-14;
-const double PI = acos(-1);
+const long double PI = acos(-1);
 // 下右上左 DRUL
 const vector<int> dx = {0, 1, 0, -1};
 const vector<int> dy = {1, 0, -1, 0};
